+++ src/limit.rs
@@ -38,22 +38,22 @@ pub enum FrequencyLimit {
     All,
     /// Only interested in frequencies `Frequency <= x`. Limit is inclusive.
     /// Supported values are `0 <= x <= Nyquist-Frequency`.
-    Min(f32),
+    Min(f64),
     /// Only interested in frequencies `x <= Frequency`. Limit is inclusive.
     /// Supported values are `0 <= x <= N`.
-    Max(f32),
+    Max(f64),
     /// Only interested in frequencies `1000 <= f <= 6777` for example. Both values are inclusive.
     /// The first value of the tuple is equivalent to [`FrequencyLimit::Min`] and the latter
     /// equivalent to [`FrequencyLimit::Max`]. Furthermore, the first value must not be
     /// bigger than the second value.
-    Range(f32, f32),
+    Range(f64, f64),
 }
 
 impl FrequencyLimit {
     /// Returns the minimum value, if any.
     #[inline]
     #[must_use]
-    pub const fn maybe_min(&self) -> Option<f32> {
+    pub const fn maybe_min(&self) -> Option<f64> {
         match self {
             Self::Min(min) => Some(*min),
             Self::Range(min, _) => Some(*min),
@@ -64,7 +64,7 @@ impl FrequencyLimit {
     /// Returns the maximum value, if any.
     #[inline]
     #[must_use]
-    pub const fn maybe_max(&self) -> Option<f32> {
+    pub const fn maybe_max(&self) -> Option<f64> {
         match self {
             Self::Max(max) => Some(*max),
             Self::Range(_, max) => Some(*max),
@@ -76,7 +76,7 @@ impl FrequencyLimit {
     /// Unwrapped version of [`Self::maybe_min`].
     #[inline]
     #[must_use]
-    pub fn min(&self) -> f32 {
+    pub fn min(&self) -> f64 {
         self.maybe_min().expect("Must contain a value!")
     }
 
@@ -84,13 +84,13 @@ impl FrequencyLimit {
     /// Unwrapped version of [`Self::maybe_max`].
     #[inline]
     #[must_use]
-    pub fn max(&self) -> f32 {
+    pub fn max(&self) -> f64 {
         self.maybe_max().expect("Must contain a value!")
     }
 
     /// Verifies that the frequency limit has sane values and takes the maximum possible
     /// frequency into account.
-    pub fn verify(&self, max_detectable_frequency: f32) -> Result<(), FrequencyLimitError> {
+    pub fn verify(&self, max_detectable_frequency: f64) -> Result<(), FrequencyLimitError> {
         match self {
             Self::All => Ok(()),
             Self::Min(x) | Self::Max(x) => {
@@ -119,13 +119,13 @@ impl FrequencyLimit {
 #[derive(Debug)]
 pub enum FrequencyLimitError {
     /// If the minimum value is below 0. Negative frequencies are not supported.
-    ValueBelowMinimum(f32),
+    ValueBelowMinimum(f64),
     /// If the maximum value is above Nyquist frequency. Nyquist-Frequency is the maximum
     /// detectable frequency.
-    ValueAboveNyquist(f32),
+    ValueAboveNyquist(f64),
     /// Either the corresponding value is below or above the minimum/maximum or the
     /// first member of the tuple is bigger than the second.
-    InvalidRange(f32, f32),
+    InvalidRange(f64, f64),
 }
 
 #[cfg(test)]
